CMSIS-DSP: Add new compilation flag to DSP_Lib_TestSuite to restore default behavior.
Default behavior is that all tests are enabled by default.
No need to define new compilation flags to enable them.

But there is a new mode which when enabled will allow to select which
tests must be run or not.

By default, this new mode is disabled and in this case all tests will
run.