Added window functions for use with Fourier transforms.
Regular functions for use with tone detection and noise estimation.
Flat-top windows for estimation of tone amplitudes.

Parameters of the windows are given.

Kaiser window missing (Bessel function not available in the library).

No f16 implementation since it is not accurate enough. It is better
to compute a f32 window and convert to f16 after.